% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_trait_data_BIEN.r
\name{build_trait_data_BIEN}
\alias{build_trait_data_BIEN}
\title{Retrieve and Clean BIEN Trait Data for a Species}
\usage{
build_trait_data_BIEN(
  species,
  traits,
  remove_outliers = FALSE,
  outlier_threshold = 3,
  author_info = FALSE,
  max_retries = 3
)
}
\arguments{
\item{species}{A character string specifying the species of interest.}

\item{traits}{A character vector specifying the trait names to retrieve from BIEN. Possible traits include:
\itemize{
\item "diameter at breast height (1.3 m) (cm)"
\item "flower color"
\item "flower pollination syndrome"
\item "fruit type"
\item "inflorescence length"
\item "leaf area (mm2)"
\item "leaf area per leaf dry mass (m2.kg-1)"
\item "leaf carbon content per leaf dry mass (mg.g-1)"
\item "leaf carbon content per leaf nitrogen content"
\item "leaf compoundness"
\item "leaf dry mass (g)"
\item "leaf dry mass per leaf fresh mass (mg.g-1)"
\item "leaf fresh mass"
\item "Leaf lamina fracture toughness"
\item "leaf life span (months)"
\item "leaf nitrogen content per leaf area (kg.m-2)"
\item "leaf nitrogen content per leaf dry mass (mg.g-1)"
\item "leaf phosphorus content per leaf area (g.m-2)"
\item "leaf phosphorus content per leaf dry mass (mg.g-1)"
\item "leaf photosynthetic rate per leaf area (µmol.m-2.s-1)"
\item "leaf photosynthetic rate per leaf dry mass (µmol.g-1.s-1)"
\item "leaf relative growth rate"
\item "leaf stomatal conductance for H2O per leaf area (mol.H2O.m-2.s-1)"
\item "leaf stomatal conductance per leaf area"
\item "leaf thickness"
\item "longest whole plant longevity (years)"
\item "maximum fruit length"
\item "maximum leaf length"
\item "maximum leaf width"
\item "maximum whole plant height"
\item "maximum whole plant longevity"
\item "minimum fruit length"
\item "minimum leaf length"
\item "minimum leaf width"
\item "minimum whole plant height"
\item "plant flowering begin (month)"
\item "plant flowering duration (month)"
\item "plant fruiting duration"
\item "root dry mass"
\item "seed length"
\item "seed mass (mg)"
\item "stem dry mass"
\item "stem relative growth rate"
\item "stem wood density (g.cm-3)"
\item "vessel lumen area (mm2)"
\item "vessel number (vessels mm-2)"
\item "whole plant dispersal syndrome"
\item "whole plant growth form"
\item "whole plant growth form diversity"
\item "whole plant height (m)"
\item "whole plant primary juvenile period length (years)"
\item "whole plant sexual system"
\item "whole plant vegetative phenology"
\item "whole plant woodiness"
}}

\item{remove_outliers}{Logical. If \code{TRUE}, removes outliers based on a specified threshold. Default is \code{FALSE}.}

\item{outlier_threshold}{Numeric. The number of standard deviations from the mean to classify a value as an outlier. Default is \code{3}.}

\item{author_info}{Logical. If \code{TRUE}, it includes authorship and contact information for data collection. Default is \code{FALSE}.}
}
\value{
A dataframe containing the cleaned trait data with columns: \code{scrubbed_species_binomial} and one column per requested trait. Optional columns include: \code{project_pi} and \code{project_pi_contacts}.
}
\description{
Fetches trait data for a given plant species from the BIEN database, cleans it, and merges multiple traits into a single dataframe.
}
\details{
BIEN pulls from a wide variety of trait databases, making it more likely to include information on a broader range of species compared to region-specific databases. For example, the LEDA trait database (Northwestern European specific) has both SLA and seed mass data for 8 total species beloning to the Acer and Pinus genera; the BIEN trait database has SLA and seed mass data for 27 species beloning to the same genera.
}
\examples{
\dontrun{
trait_df <- build_trait_data_BIEN(species = "Pinus strobus", traits = c("leaf area per leaf dry mass", "seed mass"), remove_outliers = TRUE, outlier_threshold = 2)

print(trait_df)
}
}
