% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_treemap.r
\name{build_treemap}
\alias{build_treemap}
\title{Plot a Treemap from Rarity Classification Data}
\usage{
build_treemap(data_frame, plot = TRUE)
}
\arguments{
\item{data_frame}{A dataframe containing a classifications' column. This column is used to categorize the data and must exist in the dataframe.}

\item{plot}{A logical value indicating whether to display the plot immediately (plot = FALSE). Defaults to TRUE. If TRUE the plot is stored as a ggplot object.}
}
\value{
A ggplot object representing the treemap if \code{plot} is TRUE or NULL if \code{plot} is FALSE after printing the plot.
}
\description{
This function creates a treemap visualization from a dataframe containing a 'classifications' column. It counts the occurrences of each classification and generates a ggplot-based treemap. The user can choose to display the treemap immediately (plot = FALSE) or return it as a ggplot object for further customization (plot = TRUE).
}
\examples{
# Example usage:
df <- data.frame(classifications = c("A", "B", "A", "C", "B", "A", "C"))
tree <- build_treemap(df, plot = TRUE)
tree

}
