% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_biomes.r
\name{check_biomes}
\alias{check_biomes}
\title{Analyze Species Convex Hulls Against Biome Data}
\usage{
check_biomes(convex_hulls, biome_sf = NULL)
}
\arguments{
\item{convex_hulls}{A named list of convex hulls for each species. Each element of the list should contain a list of \code{sf} polygon objects representing the convex hulls of species' distributions.}

\item{biome_sf}{An optional \code{sf} object representing the biome polygons. If not provided, the default biome shapefile (\code{biomes.shp}) from the package will be used. The default biomes are:
\itemize{
\item \strong{Humid Tropics}
\item \strong{Dry Tropics}
\item \strong{Temperate}
\item \strong{Boreal}
}}
}
\value{
A dataframe where each row corresponds to a species, and each column (except \code{species_name}) represents a biome. A value of 1 in the cell indicates that the species' convex hull intersects with the respective biome.
}
\description{
This function checks which biomes are intersected by the convex hulls of species distributions. By default, the function uses a preloaded biome shapefile containing four biomes: Humid Tropics, Dry Tropics, Temperate, and Boreal. Users can also provide their own biome data as an \code{sf} object.
}
\note{
If the CRS of the provided \code{biome_sf} or \code{convex_hulls} is not set, the function will automatically assign and transform them to EPSG:4326 (WGS84).
}
\examples{

# Generate a dataframe for Abies cephalonica and Abies firma
test_data <- data.frame(species_name = c("Abies_cephalonica", "Abies_firma"))

# Get convex hulls for species' ranges 
convex_hulls_abies <- get_range_convex_hulls(test_data)

# Check the biomes of Abies cephalonica and Abies firma - Abies cephalonica occurs in temperate forests while Abies firma occures in both temperate forests and humid tropics
abies_biomes <- check_biomes(convex_hulls_abies)

#' @references
\insertRef{Hansen2010}{GeoFunPhy}

}
