% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_eco_status.r
\name{check_eco_status}
\alias{check_eco_status}
\title{Classify Ecological Status}
\usage{
check_eco_status(data_frame, range_size_col, mean_evol_dist_col, fun_dist_col)
}
\arguments{
\item{data_frame}{A data frame containing scaled (typically using the median and median absolute deviation) values of range size, evolutionary distinctiveness, and functional distinctiveness.}

\item{range_size_col}{A column within the data frame corresponding to the range size (km^2) of the species; should be z transformed (typically using the median and meadian absolute deviation) for ecological status classification. Use the name of the column in "".}

\item{mean_evol_dist_col}{A column within the data frame corresponding to the evolutionary distinctivess of the species; should be z transformed (typically using the median and meadian absolute deviation) for ecological status classification. Use the name of the column in "".}

\item{fun_dist_col}{A column within the data frame corresponding to the functional distinctivess of the respective species. Use the name of the column in "".}
}
\value{
A data frame with species eco-evoultionary rarity classifications in the "classifications" column.
}
\description{
Categorize species in a data frame with scaled/transformed data for range size and evolutionary distinctiveness into one of eight ecological statuses.
}
\details{
The function categorizes species based on their ecological status, considering three axes of rarity: range size, evolutionary distinctiveness, and functional distinctiveness. Thresholds are set by user input, using optimal K means clustering for each variable.
}
\examples{

Create a dataframe of maple, ash, and pine species with range size, evolutionary distinctiveness, and functional distinctiveness
species_names <- c("Abies_alba", "Abies_grandis", "Abies_nordmanniana", "Acer_campestre", "Acer_monspessulanum", "Acer_negundo", "Acer_opalus", "Acer_platanoides", "Acer_pseudoplatanus", "Acer_saccharinum", "Fraxinus_angustifolia", "Fraxinus_excelsior", "Fraxinus_ornus", "Fraxinus_pennsylvanica", "Pinus_banksiana", "Pinus_cembra", "Pinus_nigra", "Pinus_pinaster", "Pinus_pinea", "Pinus_ponderosa", "Pinus_strobus", "Pinus_sylvestris", "Pinus_uncinata")

FD_values <- runif(min = -2, max = 2, n=23)

range_values <- runif(min = -2, max = 2, n=23)

evol_dist_values <- runif(min = -2, max = 2, n=23)

forest_data <- data.frame(species_name = species_names, fun_dist = FD_values, range_size = range_values, mean_evol_dist = evol_dist_values)

Find the "elbows" in the k means plots below
elbow_plot(forest_data, "range_size")
elbow_plot(forest_data, "mean_evol_dist")
elbow_plot(forest_data, "fun_dist")

eco_stat <- check_eco_status(data_frame = forest_data, fun_dist = "fun_dist", range_size = "range_size", mean_evol_dist = "mean_evol_dist")

print(eco_stat)

}
\seealso{
To determine the optimal K means to utilize here, examine the elbow plot produced using EcoStatusR::elbow_plot().

Please keep in mind that example values are randomized and therefore not representative of true ecological status.
}
\author{
Alivia G Nytko, \email{anytko@vols.utk.edu}
}
