% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_optimal_k.r
\name{choose_optimal_k}
\alias{choose_optimal_k}
\title{Choose Optimal Number of Clusters (k)}
\usage{
choose_optimal_k(variable)
}
\arguments{
\item{variable}{The name of the variable (range size, evolutionary distinctiveness, or functional distinctiveness) for which the optimal number of clusters is being chosen.}
}
\value{
An integer representing the optimal number of clusters (k).
}
\description{
Prompt the user to input the optimal number of clusters (k) for a given variable.
}
\details{
This function prompts the user to input the optimal number of clusters (k) for a given variable. The user input should be a numeric value representing the desired number of clusters. The function then saves and returns this value.
}
\examples{

Enter the chosen k value for stem length (using a method such as plotting the elbow curve)

}
