% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_trait_data_BIEN.r
\name{clean_trait_data_BIEN}
\alias{clean_trait_data_BIEN}
\title{Clean BIEN Trait Data}
\usage{
clean_trait_data_BIEN(
  data,
  remove_outliers = FALSE,
  outlier_threshold = 3,
  author_info = FALSE
)
}
\arguments{
\item{data}{A dataframe containing plant trait data. Must include the columns: \code{scrubbed_species_binomial}, \code{trait_name}, \code{trait_value}, \code{unit}, \code{method}, and \code{url_source}.}

\item{remove_outliers}{Logical. If \code{TRUE}, removes outliers based on a specified threshold. Default is \code{FALSE}.}

\item{outlier_threshold}{Numeric. The number of standard deviations from the mean to classify a value as an outlier. Default is \code{3}.}

\item{author_info}{Logical. If \code{TRUE}, it includes authorship and contact information for data collection. Default is \code{FALSE}.}
}
\value{
A cleaned dataframe with the columns: \code{scrubbed_species_binomial} and \code{mean_trait_value}. Optional columns include: \code{project_pi} and \code{project_pi_contacts}.
}
\description{
Cleans and processes plant trait data from BIEN by handling missing values, removing duplicates,computing mean trait values per species, and optionally removing outliers.
}
\examples{
\dontrun{
data <- data.frame(scrubbed_species_binomial = c("Quercus robur", "Quercus robur", "Pinus sylvestris"), trait_name = c("Leaf Area", "Leaf Area", "Needle Length"), trait_value = c(20, 22, 5), unit = c("cm2", "cm2", "cm"),method = c("measurement", "measurement", "observation"), url_source = c("source1", "source2", "source3"))

cleaned_data <- clean_trait_data_BIEN(data, remove_outliers = FALSE)

print(cleaned_data)
}
}
