% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_range_polygons.r
\name{create_range_polygons}
\alias{create_range_polygons}
\title{Create range polygons}
\usage{
create_range_polygons(
  data_frame,
  species_name = NULL,
  min_points = 5,
  min_distance = 1,
  gbif_limit = 2000
)
}
\arguments{
\item{data_frame}{A data frame containing species data.}

\item{species_name}{A character vector of species names.}

\item{min_points}{Minimum number of points to form a cluster using dbscan.}

\item{min_distance}{Minimum distance between points in the same cluster.}

\item{gbif_limit}{Maximum number of records to fetch from GBIF.}
}
\value{
A list of convex hulls for each species, or NULL if no data is found.
}
\description{
Create range polygons
}
\details{
The function creates convex hulls for species in a dataframe using gbif occurence data. Convex hulls will vary based on the minimum points and distance between points used to define a polygon.
}
\examples{
data <- data.frame(species_name = c("Quercus alba"))
convex_hulls <- create_range_polygons(data)

print(convex_hulls)

}
