% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbow_plot.r
\name{elbow_plot}
\alias{elbow_plot}
\title{Display Elbow Plot}
\usage{
elbow_plot(data, variable, k_max = 10, ggplot = FALSE)
}
\arguments{
\item{data}{A data frame containing the data for which the elbow plot is to be displayed.}

\item{variable}{The variable in the data frame for which the elbow plot is to be generated. Use the name of the column in "".}

\item{k_max}{The maximum number of clusters (k) to consider. Default is 10.}

\item{ggplot}{Logical indicator. Default is FALSE. If FALSE the function returns a R base plot. If TRUE the function returns a ggplot.}
}
\description{
Plot the Within-Cluster Sum of Squares (WCSS) against the number of clusters (k) to determine the optimal value of k using the Elbow Method.
}
\details{
The Elbow Method helps in finding the optimal number of clusters (k) by plotting the Within-Cluster Sum of Squares (WCSS) against different values of k. The "elbow" point in the plot, where the rate of decrease in WCSS slows down, indicates the optimal value of k. This is often visually determined by finding the "elbow" in the plot.
}
\examples{

Create dataframe of maple, ash, and pine species with range size, evolutionary distinctiveness, and functional distinctiveness values 
species_names <- c("Abies_alba", "Abies_grandis", "Abies_nordmanniana", "Acer_campestre", "Acer_monspessulanum", "Acer_negundo", "Acer_opalus", "Acer_platanoides", "Acer_pseudoplatanus", "Acer_saccharinum", "Fraxinus_angustifolia", "Fraxinus_excelsior", "Fraxinus_ornus", "Fraxinus_pennsylvanica", "Pinus_banksiana", "Pinus_cembra", "Pinus_nigra", "Pinus_pinaster", "Pinus_pinea", "Pinus_ponderosa", "Pinus_strobus", "Pinus_sylvestris", "Pinus_uncinata")

FD_values <- runif(min = -2, max = 2, n=23)

range_values <- runif(min = -2, max = 2, n=23)

evol_dist_values <- runif(min = -2, max = 2, n=23)

forest_data <- data.frame(species_name = species_names, fun_dist = FD_values, range_size = range_values, mean_evol_dist = evol_dist_values)

Plot the elbow plot for range size
elbow_plot(forest_data, "range_size")

Plot the elbow plot for average evolutionary distinctivess
elbow_plot(forest_data, "mean_evol_dist", 8)

Plot the elbow plot for functional distinctiveness
fun_dist_elbow <- elbow_plot(forest_data, "fun_dist", 10, ggplot = TRUE)

}
\seealso{
Use this function to visualize the optimal k value before computing eco-evolutionary statusing using GeoFunPhy::check_EER_status_k().
}
