% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restriction_performance.r
\name{eval_restriction_performance}
\alias{eval_restriction_performance}
\title{Evaluate Restriction Performance Using ANOVA}
\usage{
eval_restriction_performance(df_list, response_vars)
}
\arguments{
\item{df_list}{List of data frames each containing species data and restrictions (starting after "_flag" columns).}

\item{response_vars}{Character vector of response variable names to test in the linear models. Must be included in the df_list.}
}
\value{
A data frame combining ANOVA results for all models and response variables, with columns:
\itemize{
\item method: Identifier of the combo/data frame.
\item model: Restriction name.
\item response_variable: Response variable tested.
\item AIC: Akaike Information Criterion value for the model.
\item df: Degrees of freedom used in the ANOVA.
\item F_statistic: F statistic from the ANOVA.
\item p_value: p-value from the ANOVA F-test.
\item multiple_R2: Multiple R-squared of the model.
\item adjusted_R2: Adjusted R-squared of the model.
}
}
\description{
Applies linear models for each restriction in a list of data frames, testing ability of restrictions to explain specified response variables.
Extracts ANOVA statistics and model summaries.
}
\note{
To run highly multidimensional models such as GDR, the number of species must outnumber the possible rarity types (i.e., at least 63 species for GDR) to ensure the linear models work correctly.
}
