% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phy.r
\name{get_phy}
\alias{get_phy}
\title{Get Phylogeny}
\usage{
get_phy(phy_choice)
}
\arguments{
\item{phy_choice}{Choice of seed phylogeny includes ALLMB, ALLOTB, GBMB, GBOTB}
}
\value{
A rooted phylogenetic object to be used in comparative methods
}
\description{
Retreive seed plant phylogenies described in Smith & Brown 2018 from FePhyFoFum in github
}
\details{
The function downloads a pre-constructed phylogenetic tree from the 'big_seed_plant_trees' repository on GitHub.

The available tree choices are as follows:
\itemize{
\item 'ALLMB': GenBank and Open Tree of Life taxa with a backbone provided by Magallón et al. 2015
\item 'ALLOTB': GenBank and Open Tree of Life taxa with a backbone provided by Open Tree of Life version 9.1
\item 'GBMB': GenBank taxa with a backbone provided by Magallón et al. 2015
\item 'GBOTB': GenBank taxa with a backbone provided by Open Tree of Life version 9.1
}
}
\examples{
ALLMB_phy <- get_phy(phy_choice = "ALLMB")

ALLOTB_phy <- get_phy(phy_choice = "ALLOTB")

GMBM_phy <- get_phy(phy_choice = "GBMB")

GBOTB_phy <- get_phy(phy_choice = "GBOTB")

}
\references{
\insertRef{smithbrown2018}{EcoStatusR}

\insertRef{kleyer2008}{EcoStatusR}

\insertRef{Magallon2015}{EcoStatusR}
}
\author{
Alivia G Nytko, \email{anytko@vols.utk.edu}
}
