% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phy_angio.r
\name{get_phy_angio}
\alias{get_phy_angio}
\title{Get Phylogeny of Angiosperms}
\usage{
get_phy_angio(phy_choice)
}
\arguments{
\item{phy_choice}{Character string specifying which seed plant phylogeny to download.
Must be one of:
\itemize{
\item \code{"ALLMB"}: GenBank and Open Tree of Life taxa with a backbone from Magallón et al. (2015)
\item \code{"ALLOTB"}: GenBank and Open Tree of Life taxa with a backbone from Open Tree of Life version 9.1
\item \code{"GBMB"}: GenBank taxa with a backbone from Magallón et al. (2015)
\item \code{"GBOTB"}: GenBank taxa with a backbone from Open Tree of Life version 9.1
}}
}
\value{
A rooted phylogenetic tree as a \code{phylo} object to be used in comparative methods
}
\description{
Retreive seed plant phylogenies described in Smith & Brown 2018 from FePhyFoFum in github
}
\details{
The function downloads a pre-constructed phylogenetic tree from the
\href{https://github.com/FePhyFoFum/big_seed_plant_trees}{big_seed_plant_trees}
GitHub repository, unzips it to a temporary directory, and reads the \code{.tre} file
corresponding to the selected \code{phy_choice}.
}
\examples{
\dontrun{
ALLMB_phy <- get_phy_angio(phy_choice = "ALLMB")

ALLOTB_phy <- get_phy_angio(phy_choice = "ALLOTB")

GMBM_phy <- get_phy_angio(phy_choice = "GBMB")

GBOTB_phy <- get_phy_angio(phy_choice = "GBOTB")
}

}
\references{
\insertRef{smithbrown2018}{GDRarity}
\insertRef{kleyer2008}{GDRarity}
\insertRef{Magallon2015}{GDRarity}
}
\author{
Alivia G Nytko, \email{anytko@vols.utk.edu}
}
