% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_est_types.r
\name{map_est_types}
\alias{map_est_types}
\title{Map Rarity Types to Established Restrictions.}
\usage{
map_est_types(model, label)
}
\arguments{
\item{model}{A character string representing the restriction (e.g., "GRFRPR", "GRGLFRFL").}

\item{label}{A character string representing the presence or absence of rarity for each dimension.}
}
\value{
A character string with the rarity type if a match exists to a established restriction.
}
\description{
This function maps restriction-specific rarity types to restrictions "EER" and "Functional Rarity". It uses standardized restriction names created via \code{rename_model_code()}.
}
\examples{
map_est_types("GRFRPR", "GR+FR-PR+")
map_est_types("GRGLFRFL", "GR-GL-FR-FL-")

}
\seealso{
\code{\link[=rename_model_code]{rename_model_code()}} for how model codes are standardized.
}
