% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clipped_hulls.r
\name{plot_clipped_hulls}
\alias{plot_clipped_hulls}
\title{Plot clipped convex hulls on map}
\usage{
plot_clipped_hulls(clipped_convex_hulls)
}
\arguments{
\item{clipped_convex_hulls}{A list of clipped convex hulls for each species.}
}
\description{
Plot clipped convex hulls on map
}
\details{
This function plots a list of clipped convex hulls onto a world leaflet map.
}
\examples{
data <- data.frame(species_name = c("Eucalyptus gunnii", "Eucalyptus tenuiramis", "Eucalyptus globulus"))
continent_bounds <- get_continent_sf()
euc_convex_hulls <- get_range_convex_hulls(data)
euc_clipped_hulls <- clip_polygons_to_land(euc_convex_hulls, continent_bounds)
plot_clipped_hulls(clipped_convex_hulls = euc_clipped_hulls)

}
