% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eco_status.r
\name{plot_eco_status}
\alias{plot_eco_status}
\title{Create Interactive 3d Scatter Plot}
\usage{
plot_eco_status(data_frame, fun_dist, evol_dist, range_size)
}
\arguments{
\item{data_frame}{A data frame containing species data.}

\item{fun_dist}{A column within the data frame corresponding to the functional distinctivess of the species}

\item{evol_dist}{A column within the data frame corresponding to the evolutionary distinctivess of the species; should be z transformed (typically using the median and meadian absolute deviation) for ecological status classification.}

\item{range_size}{A column within the data frame corresponding to the range size (km^2) of the species; should be z transformed (typically using the median and meadian absolute deviation) for ecological status classification.}
}
\value{
An interactive 3d rgl object that visualizes the three dimensions of rarity and resultant ecological status. To capture a still image of the scatter plot rgl::rgl.snapshot() is recommended.
}
\description{
Create an interactive scatter plot depicting the ecological status of each species in 3d space
}
\details{
The function checks if the specified fun_dist, evol_dist, and range_size columns exists in the data frame and if the required columns 'classifications,' and 'species_name' are present. It then assigns colors to data points based on classifications and displays the 3D scatter plot interactively.
}
\seealso{
Please keep in mind that example classifications are randomized and therefore not representative of true ecological status.
}
\author{
Alivia G Nytko, \email{anytko@vols.utk.edu}
}
