% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_poly_map.r
\name{range_poly_map}
\alias{range_poly_map}
\title{Plot range polygons on world map}
\usage{
range_poly_map(
  data_frame,
  species_name = NULL,
  min_points = 5,
  min_distance = 1,
  gbif_limit = 2000,
  plot = TRUE
)
}
\arguments{
\item{species_name}{A character vector of species names.}

\item{min_points}{Minimum number of points to form a cluster using dbscan.}

\item{min_distance}{Minimum distance between points in the same cluster.}

\item{gbif_limit}{Maximum number of records to fetch from GBIF.}

\item{plot}{Logical indicating whether to plot the convex hulls on a map.}
}
\value{
A list of convex hulls for each species, or NULL if no data is found.
}
\description{
Plot range polygons on world map
}
\details{
This function combines the functionality of EcoStatusR::create_range_polygons() and EcoStatusR::plot_convex_hulls(). Specifcally this function creates convex hulls for species using occurence data and plots them onto a world leaflet map.
}
\examples{

data <- data.frame(species_name = c("Quercus alba"))

range_poly_map(data_frame = data)

}
