% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_sizes.r
\name{range_sizes}
\alias{range_sizes}
\title{Compute range sizes (sq km) from clipped or raw range polygons}
\usage{
range_sizes(clipped_polygons_list)
}
\arguments{
\item{clipped_polygons_list}{A list of clipped range polygons.}
}
\description{
This function calculates the range size (sq km) of species range polygons.
}
\examples{
# Create a dataframe with a 'species_name' column
test_data <- data.frame(species_name = c("Acer_campestre", "Acer_platanoides"))
# Generate range polygons for the speceis in the dataframe
polygon_list <- get_range_convex_hulls(test_data)
# Get continent boundaries
continent_sf <- get_continent_sf()
# Clip these range polygons to the continent bounds
clipped_polygon_list <- clip_polygons_to_land(polygon_list, continent_sf)
# Generate range sizes for each speices using the clipped range polygons
sizes <- range_sizes(clipped_polygons_list = clipped_polygon_list)
}
