% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_models.r
\name{rename_model_columns}
\alias{rename_model_columns}
\title{Rename Columns in a Data Frame Based on Restriction Names}
\usage{
rename_model_columns(df, models)
}
\arguments{
\item{df}{A data frame containing columns to rename.}

\item{models}{Character vector of restrictions (model codes corresponding to column names in \code{df}.}
}
\value{
The data frame with renamed columns where applicable.
}
\description{
This function takes a data frame and a character vector of restrictions. It renames any columns in the data frame matching the restrictions using the simplified names produced by \code{rename_model_code}.
}
\examples{
df <- data.frame(GRGL = 1:3, GRFRPR = 4:6)
rename_model_columns(df, c("GRGL", "GRFRPR"))
}
