% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_evol_dist.r
\name{avg_evol_dist}
\alias{avg_evol_dist}
\title{Calculate average evolutionary distinctiveness across multiple time scales}
\usage{
avg_evol_dist(
  phy,
  data_frame,
  taxon = NULL,
  time_slices = c(25, 50, 75, 100, 115),
  num_cores = 1
)
}
\arguments{
\item{phy}{A rooted, time-calibrated (branch lengths representative of time) phylogenetic tree used to calculate evolutionary distinctiveness.}

\item{data_frame}{A data frame containing species data with a "species_name" column.}

\item{taxon}{A character vector specifying the taxa (species) of interest. Default is NULL.}

\item{time_slices}{A numeric vector specifying the desired time slices to consider. Default is 25 MYA, 50 MYA, 75 MYA, 100MYA, 115MYA. These should be adjusted based on the time-extent of the phylogeny; the default times are based on a large phylogeny of Angiosperms. If the chosen time slices are not reflected on the utilized phylogeny, an error will occur.}

\item{num_cores}{Number of CPU cores to use for parallel processing. Default is 1. We recommend utilizing as many cores as possible.}
}
\value{
A data frame with species names and their corresponding evolutionary distinctiveness values. Species not included in the phylogeny will return NA.
}
\description{
This function calculates the average evolutionary distinctiveness for a list of species across distinct time slices.
}
\details{
Many species are evolutionarily distinct within genera but are evolutionarily indistinct across larger phylogenetic extents. This function accounts for differences in evolutionary distinctiveness at different time extents through the calculation of evolutionary distinctiveness at five distinct phylogenetic extents.
}
\examples{

Import phylogeny
GBMB_phy <- get_phy_angio(phy_choice = "GBMB")

Create dataframe with species names
maple_data <- data.frame(species_name = c("Acer_campestre", "Acer_monspessulanum", "Acer_negundo", "Acer_opalus", "Acer_platanoides", "Acer_pseudoplatanus", "Acer_saccharinum"))

Calculate evolutionary distinctiveness through time (1MYA, 5MYA, 10MYA)
GBMB_ED <- avg_evol_dist(phy = GBMB_phy, data_frame = maple_data, num_cores=6, time_slices = c(1,5,10))

print(GBMB_ED)

}
