% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_trait_data_LEDA.r
\name{build_trait_data_LEDA}
\alias{build_trait_data_LEDA}
\title{Build Trait Data from the LEDA Traitbase}
\usage{
build_trait_data_LEDA(columns_to_select, genera = NULL)
}
\arguments{
\item{columns_to_select}{A character vector specifying which traits to select. Options include: SLA (specific leaf area mm^2/mg), seed_mass (mg), leaf_mass (mg), and canopy_height (m)}

\item{genera}{A character vector specifying genera of interest to be filtered. Default is NULL.}
}
\value{
A dataframe including mean trait data for all species or species of choice in respective columns titled by trait name.
}
\description{
Build a dataframe of trait data for various plant species of interest using LEDA trait database.
}
\details{
The function reads trait data from four different data files, including SLA (Specific Leaf Area), seed mass, leaf mass, and canopy height. It then combines these data frames into a single data frame based on the species name.
}
\examples{

Build dataframe using specific leaf area (SLA) trait data across all available species
all_traits <- build_trait_data(columns_to_select = "SLA")

Build dataframe using seed mass trait data across maple species
maple_traits <- build_trait_data(columns_to_select = "seed_mass", genera = "Acer_")
print(maple_traits)

Build dataframe using canopy height for major tree species
tree_traits <- build_trait_data_LEDA(columns_to_select = "canopy_height", genera = c("Acer_", "Quercus_", "Populus_", "Ulmus_", "Pinus_", "Alnus_", "Betula_", "Salix_", "Abies_", "Fraxinus_", "Tsuga_", "Prunus_"))
print(tree_traits)

}
\references{
\insertRef{kleyer2008}{GeoFunPhy}
}
\author{
Alivia G Nytko, \email{anytko@vols.utk.edu}
}
