% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_range_size.r
\name{calc_range_size}
\alias{calc_range_size}
\title{Calculate range sizes for multiple species}
\usage{
calc_range_size(
  data_frame,
  num_cores = 1,
  min_points = 5,
  min_distance = 1,
  gbif_limit = 2000,
  continent_file = NULL
)
}
\arguments{
\item{data_frame}{A \code{data.frame} containing scienfic names of species with a column named \code{species_name}.}

\item{num_cores}{The number of cores to use for parallel processing. Default is 1 (no parallel processing).}

\item{min_points}{The minimum number of points required to form a range cluster. Default is 5.}

\item{min_distance}{The minimum distance between points in a range cluster. Default is 1 epsilon (eps).}

\item{gbif_limit}{The maximum number of GBIF records to retrieve per species. Default is 2000.}

\item{continent_file}{The URL of a GeoJSON or shape (.shp & .shx) file containing country or continent data. Default is NULL - If \code{NULL}, a GeoJSON file from Natural Earth will be used.}
}
\value{
A \code{data.frame} with two columns: \code{species_name} and \code{range_size}. The \code{range_size} column represents the size of the range for each species in square kilometers.
}
\description{
This function calculates the range sizes (sq km) of multiple species by retrieving occurrence data from GBIF, generating convex hulls, and clipping them to continent boundaries. It uses parallel processing to handle multiple species efficiently.
}
\examples{
# Generate a test dataframe with species Abies cephalonica
test_data <- data.frame(species_name = c("Abies_cephalonica"))

# Generate range sizes for each speices
range_size <- calc_range_size(data_frame = test_data)
print(range_size)
}
