% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_polygons_to_land.r
\name{clip_polygons_to_land}
\alias{clip_polygons_to_land}
\title{Clip species range polygons to land boundaries}
\usage{
clip_polygons_to_land(polygons_list, continent_sf)
}
\arguments{
\item{polygons_list}{A nested list of \code{sf} objects representing the convex hulls of a species range.}

\item{continent_sf}{A file of continent boundaries. We recommend using Natural Earth's GeoJSON or shape files.}
}
\value{
A nested list of clipped range polygons
}
\description{
This function uses a continent GeoJSON or shape file (.shp & .shx) file to clip range polygons to land boundaries.
}
\details{
When clipping the polygons of species with no points touching land (i.e. an oceanic polygon) the function will return an error. This error has been silenced and the oceanic polygon will be removed from the clipped polygon list.
}
\examples{
Generate a test dataframe with species Abies cephalonica
test_data <- data.frame(species_name = c("Abies_cephalonica"))
Retrieve the unclipped range convex hulls 
unclipped_hulls <- get_range_convex_hulls(test_data)
# Get continent data from a GeoJSON file
continent_sf_example <- get_continent_sf("https://d2ad6b4ur7yvpq.cloudfront.net/naturalearth-3.3.0/ne_50m_admin_0_countries.geojson")
# Clip the range polygons to continent bounds 
clipped_polygons <- clip_polygons_to_land(polygons_list = unclipped_hulls, continent_sf = continent_sf_example)
print(clipped_polygons)
}
