% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_continent_sf.r
\name{get_continent_sf}
\alias{get_continent_sf}
\title{Get continent spatial data from a GeoJSON file}
\usage{
get_continent_sf(url = NULL)
}
\arguments{
\item{url}{Optional URL of a GeoJSON file containing country or continent data. Default is NULL - if NULL the package uses embeded continent boundaries from Natural Earth.}
}
\value{
An \code{sf} object containing polygons representing continents. The object includes a column for the continent names and corresponding geometries.
}
\description{
This function reads a GeoJSON file from a given URL, extracts continent information, and returns an sf object with continent polygons.
}
\examples{
# Get continent data from a GeoJSON file
continent_sf <- get_continent_sf()
print(continent_sf)
}
