% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_convex_hulls.r
\name{plot_convex_hulls}
\alias{plot_convex_hulls}
\title{Plot convex hulls on map}
\usage{
plot_convex_hulls(convex_hulls)
}
\arguments{
\item{convex_hulls}{A list of convex hulls for each species.}
}
\description{
Plot convex hulls on map
}
\details{
This function plots a list of (polygon) convex hulls onto a world leaflet map.
}
\examples{
data <- data.frame(species_name = c("Quercus alba"))
oak_convex_hulls <- create_range_polygons(data)
plot_convex_hulls(convex_hulls = oak_convex_hulls)

}
