% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_poly_map.r
\name{range_poly_map}
\alias{range_poly_map}
\title{Plot range polygons on world map}
\usage{
range_poly_map(
  data_frame,
  species_name = NULL,
  min_points = 5,
  min_distance = 1,
  gbif_limit = 2000,
  plot = TRUE
)
}
\arguments{
\item{species_name}{A character vector of species names.}

\item{min_points}{The minimum number of points required to form a range cluster. Default is 5.}

\item{min_distance}{The minimum distance between points in a range cluster. Default is 1 epsilon (eps).}

\item{gbif_limit}{The maximum number of GBIF records to retrieve per species. Default is 2000.}

\item{plot}{Logical indicating whether to plot the convex hulls on a map.}
}
\value{
A leaflet map of species range polygons.
}
\description{
Plot range polygons on world map
}
\details{
This function combines the functionality of EcoStatusR::create_range_polygons() and EcoStatusR::plot_convex_hulls(). Specifcally this function creates convex hulls for species using occurence data from GBIF and plots them onto a world leaflet map.
}
\examples{

oak_data <- data.frame(species_name = c("Quercus alba", "Quercus rubra"))

range_poly_map(data_frame = oak_data)

}
