% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_by_median.r
\name{scale_by_median}
\alias{scale_by_median}
\title{Scale by Median}
\usage{
scale_by_median(data_frame, columns_chosen = NULL)
}
\arguments{
\item{data_frame}{A data frame containing the columns to transform.}

\item{columns_chosen}{A character vector specifying the column names to transform. Default is NULL.}
}
\value{
A scaled data frame with Z-scores computed using the median and medain absolute deviation (MAD) for the selected columns.
}
\description{
Transform values in a dataframe into z scores around the median. To classify species status, measures of range size and evolutionary distinctiveness must be tranformed.
}
\details{
Chosen columns are scaled using the median and median absolute deviation. This transformation is recommend for range size and evolultionary distinctiveness when classifying the ecological status of species using EcoStatusR::check_eco_status().
}
\examples{

Create dataframe of pine species with range sizes and evolutionary distinctiveness
pine_names <- c("Pinus_banksiana", "Pinus_cembra", "Pinus_nigra", "Pinus_pinaster", "Pinus_pinea", "Pinus_ponderosa", "Pinus_strobus", "Pinus_sylvestris", "Pinus_uncinata")
evol_dist_values <- runif(min = -50, max = 50, n=9)
range_values <- runif(min = -20000, max = 20000, n=9)

pine_data <- data.frame(species_name = pine_names, range_size = range_values, evol_dist = evol_dist_values)

transform_all <- scale_by_median(data_frame = pine_data)
print(transform_all)

}
\seealso{
Please keep in mind that these values in the examples are randomized and therefore not representative of true ecological status.
}
\author{
Alivia G Nytko, \email{anytko@vols.utk.edu}
}
